Add linear time greedy part two solution
Uses a linked list to maintain a list of digits, with constant time
deletion.

At each step, it adds the next value to the end of the list, and updates
the cache of potentially out of order nodes, then it removes the first
out of order node, or removes the last inserted.

A node is out of order if the node following it has a greater value. In
this way, we always remove the value with the largest increase to the
current value