Fix day 9 part 2 solution
Handle weird shapes
Don't try to determine every single green space, just calculate as needed.
Sadly this is still slow, doing n**2 checks across the red tiles, and then doing checks for the border of the rectangle, which run across width nodes to check if the point is in the perimiter.